requirements.txt
fastapi
uvicorn
sentence-transformers
pydantic
python-dotenv

google-generativeai
